tests/wpt/web-platform-tests/css/css-typed-om/styleMap-update-function.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>styleMap objects provide an 'update' function</title>
<meta name="author" title="Shane Stephens">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <div id='element'></div>
  <div id="log"></div>
  <script>
    test(function() {
      element.styleMap.set('width', new CSSUnitValue(42, 'px'));
      element.styleMap.update('width', length => new CSSSimpleLength(length.value + 20, length.type))
      assert_equals(element.styleMap.get('width').value, 62, 'update expected to apply callback to old value in map');
    });
  </script>
