third_party/WebKit/LayoutTests/external/wpt/css/css-typed-om/styleMap-update-function-expected.txt

This is a testharness.js-based test.
FAIL styleMap objects provide an 'update' function assert_equals: update expected to apply callback to old value in map expected 62 but got 42
Harness: the test ran to completion.
